Move iota tests to test_basic.py
